folder for aggregatedanalyses
