+++ bika/lims/browser/aggregatedanalyses/__init__.py
@@ -1,0 +1,9 @@
+# This file is part of Bika LIMS
+#
+# Copyright 2011-2016 by it's authors.
+# Some rights reserved. See LICENSE.txt, AUTHORS.txt.
+
+from .aggregatedanalyses import AggregatedAnalysesView
+from .aggregatedanalyses_filter_bar\
+    import AggregatedanalysesBikaListingFilterBar
+from .workflow import AggregatedAnalysesWorkflowAction
