+++ bika/lims/browser/aggregatedanalyses/aggregatedanalyses.py
@@ -5,33 +5,14 @@
 # Copyright 2011-2016 by it's authors.
 # Some rights reserved. See LICENSE.txt, AUTHORS.txt.
 
-from AccessControl import getSecurityManager
-from Products.CMFPlone.utils import safe_unicode
 from bika.lims import bikaMessageFactory as _
-from bika.lims.utils import t, dicts_to_dict, format_supsub
-from bika.lims.utils.analysis import format_uncertainty
-from bika.lims.browser import BrowserView
 from bika.lims.browser.analyses import AnalysesView
-from bika.lims.browser.bika_listing import BikaListingView ####
-from bika.lims.config import QCANALYSIS_TYPES
-from bika.lims.interfaces import IResultOutOfRange
 from bika.lims.permissions import *
-from bika.lims.utils import isActive
-from bika.lims.utils import getUsers
-from bika.lims.utils import to_utf8
-from bika.lims.utils import formatDecimalMark
-from DateTime import DateTime
-from operator import itemgetter
-from Products.Archetypes.config import REFERENCE_CATALOG
-from Products.CMFCore.utils import getToolByName
-from Products.CMFCore.WorkflowCore import WorkflowException
-from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
-from bika.lims.utils.analysis import format_numeric_result
-from zope.interface import implements
-from zope.interface import Interface
-from zope.component import getAdapters
-
+from bika.lims.browser.aggregatedanalyses.aggregatedanalyses_filter_bar\
+    import AggregatedanalysesBikaListingFilterBar
 import json
+from Products.CMFCore.utils import getToolByName
+from zope.interface import implements
 
 
 class AggregatedAnalysesView(AnalysesView):
@@ -63,6 +44,9 @@ class AggregatedAnalysesView(AnalysesView):
         self.form_id = 'analyses_form'
         self.portal = getToolByName(context, 'portal_url').getPortalObject()
         self.portal_url = self.portal.absolute_url()
+        # Check if the filter bar functionality is activated or not
+        self.filter_bar_enabled =\
+            self.context.bika_setup.getDisplayAdvancedFilterBarForAnalyses()
 
         # each editable item needs it's own allow_edit
         # which is a list of field names.
@@ -78,11 +62,12 @@ class AggregatedAnalysesView(AnalysesView):
             }
         self.review_states = [
             {'id': 'default',
-             'title':  _('All'),
+             'title':  _('Results pending'),
              'transitions': [{'id': 'sample'},
                              {'id': 'submit'},
                              {'id': 'cancel'},
-                             {'id': 'assign'}],
+                             # {'id': 'assign'}
+                             ],
              'contentFilter': {'review_state': [
                 'sample_received', 'assigned', 'attachment_due']},
              'columns': ['AnalysisRequest',
@@ -97,9 +82,31 @@ class AggregatedAnalysesView(AnalysesView):
                          'state_title',
                          ]
              },
+             {'id': 'to_be_verified',
+              'title':  _('To be verified'),
+              'transitions': [{'id': 'verify'},
+                              {'id': 'cancel'}
+                              ],
+              'contentFilter': {'review_state': [
+                 'to_be_verified']},
+              'columns': ['AnalysisRequest',
+                          'Worksheet',
+                          'Service',
+                          'Result',
+                          'Uncertainty',
+                          'Partition',
+                          'Method',
+                          'Instrument',
+                          'Analyst',
+                          'state_title',
+                          ]
+              },
         ]
         if not context.bika_setup.getShowPartitions():
             self.review_states[0]['columns'].remove('Partition')
+
+    def getPOSTAction(self):
+        return 'aggregatedanalyses_workflow_action'
 
     def isItemAllowed(self, obj):
         """
@@ -115,6 +122,8 @@ class AggregatedAnalysesView(AnalysesView):
         if not self.context.bika_setup.getAllowDepartmentFiltering():
             return True
         # Gettin the department from analysis service
+        if self.filter_bar_enabled and not self.filter_bar_check_item(obj):
+            return False
         serv_dep = obj.getService().getDepartment()
         result = True
         if serv_dep:
@@ -140,3 +149,14 @@ class AggregatedAnalysesView(AnalysesView):
             anchor = '<a href="%s">%s</a>' % (ws.absolute_url(), ws.Title())
             item['replace']['Worksheet'] = anchor
         return item
+
+    def getFilterBar(self):
+        """
+        This function creates an instance of BikaListingFilterBar if the
+        class has not created one yet.
+        :return: a BikaListingFilterBar instance
+        """
+        self._advfilterbar = self._advfilterbar if self._advfilterbar else \
+            AggregatedanalysesBikaListingFilterBar(
+                context=self.context, request=self.request)
+        return self._advfilterbar
