+++ bika/lims/browser/dashboard/configure.zcml
@@ -10,19 +10,5 @@
         permission="zope.Public"
         layer="bika.lims.interfaces.IBikaLIMS"
     />
-    <browser:page
-      for="*"
-      name="aggregatedanalyses"
-      class="bika.lims.browser.dashboard.AggregatedAnalysesView"
-      permission="zope.Public"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
 
-    <browser:page
-      for="*"
-      name="aggregatedanalyses_workflow_action"
-      class="bika.lims.browser.dashboard.AggregatedAnalysesWorkflowAction"
-      permission="zope.Public"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
 </configure>
